Restructure, can't use generics
Go is so restricting...
