+++ go-packages/sim2/coordmath.go
@@ -1,4 +1,4 @@
-package simulator2
+package sim2
 
 import (
   "math"
